research: 목소리 후보 셋을 같은 조건에 세우고, 셋 다 떨어뜨리다
89.4분으로 Seed-VC를 더 학습하는 반복을 여기서 멈춘다. 데이터가 모자란 것이
아니라 F0를 입력받지 않는 변환 경로가 Chatterbox의 억양을 다시 만들고 있었다.
학습된 22k 모델은 화자 유사도 0.581을 내면서 질문/평서 종결 차이는 37%,
문장 내부 억양 폭은 43%만 남긴다.

공정하게 세우는 데 대부분의 노력이 들어갔다. 화자 유사도를 합성에 쓴 ref-J로
재면 그 참조를 조건으로 쓴 모델이 그냥 이긴다. held-out ref-A~ref-I 아홉 개의
평균 embedding으로 바꿨고, 억양은 절대값이 아니라 같은 생성분에서의 보존율로,
내용은 local Whisper 재전사로 봤다. seed는 양쪽 다 고정했다.

Chatterbox의 Reference 조건이 하나가 아니라는 것도 여기서 갈라졌다. S3Gen만
Reference로 주면 0.611로 가장 높고, T3까지 주면 질문/평서 방향이 -17%로
뒤집힌다.

객관 지표만 보면 Seed-VC 44k F0 + auto_f0_adjust가 질문/평서 122%, 억양 폭
105%로 유일한 작업 후보였다. 그런데 blind 청취에서 거친 소리가 가장 심해
천우가 탈락시켰다. 공개 RVC 체크포인트도 같은 조건에 세웠으나 유사도 0.236,
질문/평서 -12%로 기준선에 못 미쳤다 -- 다만 노래용 모델이라 이 수치만으로
탈락시키지는 않는다.

운영 경로는 바꾸지 않는다. 가장 나았던 Chatterbox S3Gen Reference를 다음
모델 비교의 기준선으로 남긴다.

+++ experiments/training_set_audit.py
@@ -104,7 +104,7 @@ def widest_gap(scores: list[float]) -> float:
 
 def seconds(path: Path) -> float:
     info = torchaudio.info(str(path))
-    return info.num_frames / info.sample_rate
+    return float(info.num_frames / info.sample_rate)
 
 
 def build_parser() -> argparse.ArgumentParser:
